# library/src/main/java/com/org/library/exception/ErrorResponse.java
package com.org.library.exception;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Builder;
import lombok.Data;

import java.time.LocalDateTime;
import java.util.Map;

@Data
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ErrorResponse {

    private int                 status;
    private String              error;
    private String              message;
    private LocalDateTime       timestamp;

    // Only present on validation failures — field -> error message
    private Map<String, String> validationErrors;
}